requirements.txt
torch>=1.12.0
networkx>=2.8
pyvis>=0.3.0
pubchempy>=1.0.4
pandas>=1.4.0
rdkit-pypi>=2022.3.0
torchmeta>=1.8.0
pytest>=7.0.0
pytest-cov>=3.0.0
